requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.0.4
dj-database-url==0.4.2
Django==1.8.5
djangorestframework==3.2.4
gunicorn==19.7.1
idna==2.6
ipython-genutils==0.2.0
path.py==8.1.1
pickleshare==0.5
pyreadline==2.1
simplegeneric==0.8.1
traitlets==4.0.0
urllib3==1.22
virtualenv==13.1.2
virtualenvwrapper-win==1.2.1
whitenoise==3.3.0
psycopg2==2.4.5
coverage==3.6